confirm dialog for cancelling booking

+++ lib/ui/pages/reservation/upcoming_list.dart
@@ -44,25 +44,34 @@ class UpcomingList extends StatelessWidget {
                       const Expanded(child: Text("")),
                       TextButton(
                         onPressed: () async {
-                          try {
-                            await Supabase.instance.client
-                                .from("reservations")
-                                .delete()
-                                .eq("id", reservation['id']);
+                          showDialog(
+                            context: context,
+                            builder: (BuildContext context) {
+                              return const CancelBookingDialog();
+                            },
+                          ).then((confirmed) async {
+                            if (confirmed != null && confirmed) {
+                              try {
+                                await Supabase.instance.client
+                                    .from("reservations")
+                                    .delete()
+                                    .eq("id", reservation['id']);
 
-                            // revalidate the data
-                            refreshReservation();
+                                // revalidate the data
+                                refreshReservation();
 
-                            if (!context.mounted) return;
-                            showKwunSnackBar(
-                                context: context,
-                                message: "Your booking has been cancelled",
-                                color: Colors.green);
-                          } on Exception catch (e) {
-                            if (!context.mounted) return;
-                            showKwunSnackBar(
-                                context: context, message: e.toString());
-                          }
+                                if (!context.mounted) return;
+                                showKwunSnackBar(
+                                    context: context,
+                                    message: "Your booking has been cancelled",
+                                    color: Colors.green);
+                              } on Exception catch (e) {
+                                if (!context.mounted) return;
+                                showKwunSnackBar(
+                                    context: context, message: e.toString());
+                              }
+                            }
+                          });
                         },
                         child: Container(
                             padding: const EdgeInsets.all(5),
@@ -85,6 +94,52 @@ class UpcomingList extends StatelessWidget {
           );
         },
       ),
+    );
+  }
+}
+
+class CancelBookingDialog extends StatelessWidget {
+  const CancelBookingDialog({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return AlertDialog(
+      title: const Text('Cancel Booking'),
+      content: const Text('Are you sure you want to cancel this booking?'),
+      actions: <Widget>[
+        TextButton(
+          onPressed: () {
+            Navigator.of(context)
+                .pop(false); // Return false when cancel is pressed
+          },
+          child: Container(
+              decoration: BoxDecoration(
+                  color: Colors.black26,
+                  borderRadius: BorderRadius.circular(20)),
+              child: const Padding(
+                padding: EdgeInsets.only(
+                    left: 20.0, right: 20.0, top: 8.0, bottom: 8.0),
+                child: Text(
+                  'No',
+                ),
+              )),
+        ),
+        TextButton(
+          onPressed: () {
+            Navigator.of(context)
+                .pop(true); // Return true when confirm is pressed
+          },
+          child: Container(
+              decoration: BoxDecoration(
+                  color: Colors.black26,
+                  borderRadius: BorderRadius.circular(20)),
+              child: const Padding(
+                padding: EdgeInsets.only(
+                    left: 18.0, right: 18.0, top: 8.0, bottom: 8.0),
+                child: Text('Yes'),
+              )),
+        ),
+      ],
     );
   }
 }
